feat: Tests of first implementation

+++ src/main/java/ECIEXPRESS/AmaterasuPagos/Payment/BackEnd/Infrastructure/Clients/Promotion/Dto/PromotionResponses/PromotionResponse.java
@@ -3,6 +3,6 @@ package ECIEXPRESS.AmaterasuPagos.Payment.BackEnd.Infrastructure.Clients.Promoti
 import java.util.List;
 
 public record PromotionResponse(
-        Integer finalAmount,
+        double finalAmount,
         List<String> appliedPromotions) {
 }
